# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Fair Drop API",
    "version": "0.1.0"
  },
  "paths": {
    "/api/v1/auth/me": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Get current user identity",
        "description": "Resolves caller identity via Supabase JWT, triggers JIT provisioning on first call, and returns profile + participant state.",
        "operationId": "get_me_api_v1_auth_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrentUserResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/auth/me/admin": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Verify admin access",
        "description": "Returns caller profile if they hold ADMIN role in the profiles table, otherwise returns 403 Forbidden.",
        "operationId": "get_me_admin_api_v1_auth_me_admin_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/campaigns": {
      "get": {
        "tags": [
          "campaigns"
        ],
        "summary": "List published campaigns",
        "description": "Returns a paginated list of non-draft campaigns. Supports optional status filtering.",
        "operationId": "list_campaigns_api_v1_campaigns_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/campaigns/{campaign_id}/status": {
      "get": {
        "tags": [
          "campaigns"
        ],
        "summary": "Get participant campaign status (recovery endpoint)",
        "description": "Returns current participant registration state, admission availability, and campaign lifecycle state from server truth.",
        "operationId": "get_campaign_status_endpoint_api_v1_campaigns__campaign_id__status_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "campaign_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Campaign Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/campaigns/{campaign_id}/result": {
      "get": {
        "tags": [
          "campaigns"
        ],
        "summary": "Get participant lottery draw result",
        "description": "Returns personal draw result: WON (with rank and entitlement), WAITLISTED, or NOT_REGISTERED.",
        "operationId": "get_participant_result_endpoint_api_v1_campaigns__campaign_id__result_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "campaign_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Campaign Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParticipantResultResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/campaigns/{campaign_id}/join": {
      "post": {
        "tags": [
          "registration"
        ],
        "summary": "Join campaign waiting room and obtain cryptographic admission permit",
        "description": "Validates campaign state and issues a short-lived, HMAC-SHA256 signed admission permit token.",
        "operationId": "join_campaign_endpoint_api_v1_campaigns__campaign_id__join_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "campaign_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Campaign Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/JoinRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JoinResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/campaigns/{campaign_id}/register": {
      "post": {
        "tags": [
          "registration"
        ],
        "summary": "Register participant entry for campaign with idempotency",
        "description": "Atomically consumes the admission permit, evaluates risk/challenge seams, and creates a unique registration entry.",
        "operationId": "register_campaign_endpoint_api_v1_campaigns__campaign_id__register_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "campaign_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Campaign Id"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CampaignListResponse": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/CampaignPublicResponse"
            },
            "type": "array",
            "title": "Data"
          },
          "meta": {
            "additionalProperties": true,
            "type": "object",
            "title": "Meta",
            "description": "Pagination metadata: page, page_size, total"
          }
        },
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "title": "CampaignListResponse",
        "description": "Paginated collection of public campaign records."
      },
...
      "CampaignStatusResponse": {
        "properties": {
          "campaign": {
            "additionalProperties": true,
            "type": "object",
            "title": "Campaign"
          },
          "participant_state": {
            "type": "string",
            "title": "Participant State",
            "description": "One of: NOT_REGISTERED, REGISTERED, SELECTED, STANDBY, etc."
          },
          "registration": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Registration"
          },
          "registration_slice": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Registration Slice"
          },
          "admission": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Admission"
          },
          "challenge": {
            "type": "null",
            "title": "Challenge"
          },
          "entitlement": {
            "type": "null",
            "title": "Entitlement"
          },
          "seat_hold": {
            "type": "null",
            "title": "Seat Hold"
          }
        },
        "type": "object",
        "required": [
          "campaign",
          "participant_state"
        ],
        "title": "CampaignStatusResponse",
        "description": "Client recovery endpoint response returning server truth for a participant."
      },
...
      "CurrentUserResponse": {
        "properties": {
          "profile": {
            "$ref": "#/components/schemas/ProfileResponse"
          },
          "participant": {
            "$ref": "#/components/schemas/ParticipantResponse"
          },
          "server_time": {
            "type": "string",
            "format": "date-time",
            "title": "Server Time",
            "description": "Server timestamp in UTC"
          }
        },
        "type": "object",
        "required": [
          "profile",
          "participant",
          "server_time"
        ],
        "title": "CurrentUserResponse",
        "description": "Combined response returned by /auth/me representing the authenticated identity."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "JoinRequest": {
        "properties": {
          "client_meta": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Meta",
            "description": "Optional client telemetry or environment context."
          }
        },
        "type": "object",
        "title": "JoinRequest",
        "description": "Optional payload for requesting an admission permit."
      },
      "JoinResponse": {
        "properties": {
          "permit_id": {
            "type": "string",
            "format": "uuid",
            "title": "Permit Id"
          },
          "admission_token": {
            "type": "string",
            "title": "Admission Token",
            "description": "Signed cryptographic admission permit token to present on registration."
          },
          "nonce": {
            "type": "string",
            "title": "Nonce",
            "description": "Cryptographically secure 32-character single-use nonce."
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          },
          "campaign_id": {
            "type": "string",
            "format": "uuid",
            "title": "Campaign Id"
          },
          "server_time": {
            "type": "string",
            "format": "date-time",
            "title": "Server Time"
          }
        },
        "type": "object",
        "required": [
          "permit_id",
          "admission_token",
          "nonce",
          "expires_at",
          "campaign_id",
          "server_time"
        ],
        "title": "JoinResponse",
        "description": "Response returned upon joining a campaign waiting room."
      },
...
      "ParticipantResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id",
            "description": "Participant UUID"
          },
          "account_id": {
            "type": "string",
            "format": "uuid",
            "title": "Account Id",
            "description": "Linked Profile UUID"
          },
          "verification_status": {
            "type": "string",
            "enum": [
              "PENDING",
              "VERIFIED",
              "REJECTED"
            ],
            "title": "Verification Status",
            "description": "Fair Drop participant verification status"
          },
          "risk_level": {
            "type": "string",
            "enum": [
              "LOW",
              "MEDIUM",
              "HIGH"
            ],
            "title": "Risk Level",
            "description": "Assigned participant risk score level"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "Participant record creation timestamp"
          }
        },
        "type": "object",
        "required": [
          "id",
          "account_id",
          "verification_status",
          "risk_level",
          "created_at"
        ],
        "title": "ParticipantResponse",
        "description": "Participant risk and verification status representation."
      },
      "ParticipantResultResponse": {
        "properties": {
          "campaign_id": {
            "type": "string",
            "format": "uuid",
            "title": "Campaign Id"
          },
          "participant_id": {
            "type": "string",
            "format": "uuid",
            "title": "Participant Id"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "Result status: 'WON', 'WAITLISTED', or 'NOT_REGISTERED'."
          },
          "is_winner": {
            "type": "boolean",
            "title": "Is Winner",
            "description": "Whether participant secured a winning allocation."
          },
          "rank": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rank",
            "description": "Assigned random draw position (1-indexed)."
          },
          "entitlement_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Entitlement Id",
            "description": "ID of generated entitlement if winner."
          },
          "entitlement_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Entitlement Status",
            "description": "Status of entitlement (e.g. SELECTED)."
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At",
            "description": "Expiration time window for claim/purchase."
          },
          "randomness_commitment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Randomness Commitment",
            "description": "SHA-256 commitment hash of draw seed."
          },
          "draw_executed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Draw Executed At",
            "description": "Timestamp when lottery draw was finalized."
          }
        },
        "type": "object",
        "required": [
          "campaign_id",
          "participant_id",
          "status",
          "is_winner"
        ],
        "title": "ParticipantResultResponse",
        "description": "Participant query response for lottery outcome and entitlement status."
      },
...
      "ProfileResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id",
            "description": "Supabase user UUID"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name",
            "description": "User display name"
          },
          "role": {
            "type": "string",
            "enum": [
              "USER",
              "ADMIN"
            ],
            "title": "Role",
            "description": "Platform authorization role"
          },
          "email_verified": {
            "type": "boolean",
            "title": "Email Verified",
            "description": "Email verification status"
          },
          "phone_verified": {
            "type": "boolean",
            "title": "Phone Verified",
            "description": "Phone verification status"
          },
          "institute_verified": {
            "type": "boolean",
            "title": "Institute Verified",
            "description": "Institute verification status"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "Account creation timestamp"
          }
        },
        "type": "object",
        "required": [
          "id",
          "role",
          "email_verified",
          "phone_verified",
          "institute_verified",
          "created_at"
        ],
        "title": "ProfileResponse",
        "description": "Public profile representation."
      },
      "RegisterRequest": {
        "properties": {
          "admission_token": {
            "type": "string",
            "minLength": 1,
            "title": "Admission Token",
            "description": "Signed admission token received from /join."
          },
          "nonce": {
            "type": "string",
            "minLength": 16,
            "title": "Nonce",
            "description": "Cryptographic nonce matching the issued admission permit."
          },
          "client_meta": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Meta",
            "description": "Optional client telemetry/device metadata."
          }
        },
        "type": "object",
        "required": [
          "admission_token",
          "nonce"
        ],
        "title": "RegisterRequest",
        "description": "Payload submitted to complete registration."
      },
      "RegisterResponse": {
        "properties": {
          "registration_id": {
            "type": "string",
            "format": "uuid",
            "title": "Registration Id"
          },
          "campaign_id": {
            "type": "string",
            "format": "uuid",
            "title": "Campaign Id"
          },
          "participant_id": {
            "type": "string",
            "format": "uuid",
            "title": "Participant Id"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "Registration lifecycle status (e.g. REGISTERED).",
            "default": "REGISTERED"
          },
          "registered_at": {
            "type": "string",
            "format": "date-time",
            "title": "Registered At"
          },
          "risk_level": {
            "type": "string",
            "title": "Risk Level",
            "description": "Evaluated risk tier (LOW, MEDIUM, HIGH).",
            "default": "LOW"
          },
          "challenge_required": {
            "type": "boolean",
            "title": "Challenge Required",
            "description": "Whether an additional proof-of-human challenge is required.",
            "default": false
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Human-readable status summary.",
            "default": "Registration accepted successfully"
          }
        },
        "type": "object",
        "required": [
          "registration_id",
          "campaign_id",
          "participant_id",
          "registered_at"
        ],
        "title": "RegisterResponse",
        "description": "Authoritative confirmation of a successfully registered participant entry."
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}